pyproject.toml
[project]
name = "agh-community-bot"
version = "0.1.0"
license = { file = "LICENSE" }
requires-python = ">=3.5"
dependencies = [
    "asyncpg>=0.23.0",
    "python-dotenv>=0.20.0",
    "sqlalchemy>=1.4.54",
    "telethon>=1.39.0",
]

[project.optional-dependencies]
speed = [
    "cryptg>=0.3.1",
]

[dependency-groups]
dev = [
    "mypy>=0.910",
    "ruff>=0.0.17",
    "types-redis>=4.4.0.0",
]

[tool.mypy]
follow_untyped_imports = true
strict = true
untyped_calls_exclude = ["telethon"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
